% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_ls.R
\name{me_ls_iid}
\alias{me_ls_iid}
\title{Compute ME-LS Estimates and Cluster-Robust Standard Errors (iid)}
\usage{
me_ls_iid(
  X,
  Y,
  Sigma,
  cluster1 = NULL,
  cluster2 = NULL,
  eigfix = FALSE,
  tol = 1e-08,
  level = 0.95
)
}
\arguments{
\item{X}{Matrix of covariates}

\item{Y}{Vector or matrix of outcomes}

\item{Sigma}{Covariance matrix for ME (for non-MARS first-steps, which have
no ME, entries will be zero)}

\item{cluster1}{Vector specifying primary clustering variable (default NULL)}

\item{cluster2}{Optional vector specifying secondary clustering variable for
two-way clustering}

\item{eigfix}{Logical indicating whether to apply eigenvalue correction
(default FALSE)}

\item{tol}{Tolerance level for eigenvalue correction (default 1e-8)}

\item{level}{Level for the confidence interval (default: 0.95)}
}
\value{
A list containing:
\itemize{
\item coefficients: Point estimates
\item vcov: Variance-covariance matrix
\item robust_se: Cluster-robust standard errors
\item p_values: Two-sided p-values
\item ci_lower: Lower bounds of confidence intervals
\item ci_upper: Upper bounds of confidence intervals
}
}
\description{
This function calculates iid classical measurement error (ME) corrected
least squares (LS) estimates along with (up to two-way) cluster-robust
standard errors.
}
