% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{residualize}
\alias{residualize}
\title{Residualize with a Linear Projection}
\usage{
residualize(target_vec, featur_mat)
}
\arguments{
\item{target_vec}{Vector of the target variable to be residualized}

\item{featur_mat}{Matrix of the features for residualization}
}
\value{
Vector of residuals from the regression
}
\description{
Computes residuals from a linear projection of the target variable on the feature matrix.
}
\keyword{internal}
